Flatten native association enums
