+++ src/WinPrint.WinForms/AssocStr.cs
@@ -1,0 +1,18 @@
+using System;
+using System.Diagnostics.CodeAnalysis;
+
+namespace WinPrint.WinForms;
+
+public enum AssocStr
+{
+    Command = 1,
+    Executable,
+    FriendlyDocName,
+    FriendlyAppName,
+    NoOpen,
+    ShellNewValue,
+    DDECommand,
+    DDEIfExec,
+    DDEApplication,
+    DDETopic
+}
